feat(agent): Complete Task 7 - Agent Orchestrator - Refactor tools into separate files
