+++ scripts/run_local_agent.py
@@ -10,54 +10,40 @@ import litellm
 from dotenv import load_dotenv
 from git_manager import GitManager
 
+# Import the refactored tools
+from scripts.tools.write_file import write_file_tool
+from scripts.tools.retrieve_context_files import retrieve_context_files_tool
+from scripts.tools.rename_files import rename_files as rename_files_tool
+from scripts.tools.submit_for_review import submit_for_review_tool
+from scripts.tools.ask_question import ask_question_tool
+from scripts.tools.finish import finish_tool
+
 # --- All classes up to UnifiedEngine are unchanged ---
 class AgentTools:
     def __init__(self, repo_path: str, git_manager: GitManager):
         self.repo_path = repo_path
         self.git_manager = git_manager
+
     def write_file(self, path: str, content: str):
-        full_path = os.path.join(self.repo_path, path)
-        os.makedirs(os.path.dirname(full_path), exist_ok=True)
-        with open(full_path, "w") as f: f.write(content)
-        return f"Successfully wrote {len(content)} bytes to {path}"
+        return write_file_tool(self.repo_path, path, content)
+
     def retrieve_context_files(self, paths: list):
-        content = {}
-        for path in paths:
-            full_path = os.path.join(self.repo_path, path)
-            try:
-                with open(full_path, "r") as f:
-                    content[path] = f.read()
-            except FileNotFoundError:
-                content[path] = f"Error: File '{path}' not found."
-            except Exception as e:
-                content[path] = f"Error reading file '{path}': {str(e)}"
-        return json.dumps(content)
+        return retrieve_context_files_tool(self.repo_path, paths)
+
     def rename_files(self, operations: list, overwrite: bool = False, dry_run: bool = False):
-        """
-        Rename or move files/directories within the repository safely.
-        - operations: list of {"from_path": str, "to_path": str}
-        - overwrite: bool (default False)
-        - dry_run: bool (default False)
-        Returns a JSON string summarizing results.
-        """
         try:
-            from rename_files import rename_files as _do_renames
+            return json.dumps(rename_files_tool(operations=operations, base_dir=self.repo_path, overwrite=overwrite, dry_run=dry_run))
         except Exception as e:
-            return json.dumps({"ok": False, "error": f"Failed to import rename_files module: {e}"})
-        result = _do_renames(operations=operations, base_dir=self.repo_path, overwrite=overwrite, dry_run=dry_run)
-        return json.dumps(result)
+            return json.dumps({"ok": False, "error": f"Failed to execute rename_files tool: {e}"})
+
     def submit_for_review(self, task_id: int, task_title: str):
-        print("Submitting changes for review...")
-        commit_message = f"feat(agent): Complete Task {task_id} - {task_title}"
-        pr_body = f"This PR was automatically generated by the agent to complete Task #{task_id}."
-        if self.git_manager.commit_and_push(commit_message):
-            if self.git_manager.create_pull_request(commit_message, pr_body):
-                return f"Successfully submitted for review. Pull Request for Task {task_id} created."
-        return "Failed to submit for review."
+        return submit_for_review_tool(self.git_manager, task_id, task_title)
+
     def ask_question(self, question_text: str):
-        return f"HALT: Agent has a question: {question_text}"
+        return ask_question_tool(question_text)
+
     def finish(self, reason: str):
-        return f"Agent finished cycle. Reason: {reason}"
+        return finish_tool(reason)
 
 class UnifiedEngine:
     def generate_plan_and_tool_calls(self, model: str, context: dict) -> list:
@@ -139,7 +125,7 @@ class Agent:
         self.engine = UnifiedEngine()
         print(f"Agent initialized. Mode: {self.mode}, Model: {self.model}. Running in Safe Mode.")
 
-    def run(self): # Unchanged
+    def run(self):
         run_count = 0
         while True:
             run_count += 1
@@ -149,7 +135,7 @@ class Agent:
                 break
         print("\n--- Agent has finished all work. ---")
 
-    def _execute_cycle(self, run_count: int) -> bool: # Unchanged
+    def _execute_cycle(self, run_count: int) -> bool:
         repo_url = self._get_repo_url()
         git_manager = GitManager(repo_url=repo_url)
         if not git_manager.setup_repository(branch_name=f"agent/cycle-{run_count}"):
@@ -199,7 +185,7 @@ class Agent:
             "docs/TASK_FORMAT.md",
             "docs/TOOL_ARCHITECTURE.md",
             "scripts/run_local_agent.py",
-            "scripts/rename_files.py",
+            "scripts/tools/rename_files.py", # Updated path
             "tasks/7/plan_7.md"
         ]
         context = {}
