+++ scripts/tools/ask_question.py
@@ -1,0 +1,2 @@
+def ask_question_tool(question_text: str):
+    return f"HALT: Agent has a question: {question_text}"
