+++ scripts/tools/finish.py
@@ -1,0 +1,2 @@
+def finish_tool(reason: str):
+    return f"Agent finished cycle. Reason: {reason}"
