+++ scripts/tools/submit_for_review.py
@@ -1,0 +1,8 @@
+def submit_for_review_tool(git_manager, task_id: int, task_title: str):
+    print("Submitting changes for review...")
+    commit_message = f"feat(agent): Complete Task {task_id} - {task_title}"
+    pr_body = f"This PR was automatically generated by the agent to complete Task #{task_id}."
+    if git_manager.commit_and_push(commit_message):
+        if git_manager.create_pull_request(commit_message, pr_body):
+            return f"Successfully submitted for review. Pull Request for Task {task_id} created."
+    return "Failed to submit for review."
